Prueba

+++ src/controller/post-controller.js
@@ -1,11 +1,13 @@
+/* eslint-disable no-alert */
+/* eslint-disable no-console */
 import {
   addPost,
   updatePost,
   deletePost,
-  addComment,
-  countLikes,
   updateTypePost,
 } from '../model/user-post.js';
+
+import { addLikes, getUserLike, deleteLikes } from '../model/likes-post.js';
 import { currentUser } from '../model/auth-user.js';
 
 export const addDataPost = (event) => {
@@ -13,7 +15,16 @@ export const addDataPost = (event) => {
   const btnShare = event.target;
   const newPost = btnShare.closest('.card-new-post').querySelector('textarea');
   const typePost = btnShare.closest('.card-new-post').querySelector('select');
-  addPost(newPost.value, typePost.value)
+  const objectPost = {
+    post: newPost.value,
+    idUser: currentUser().uid,
+    name: currentUser().displayName,
+    email: currentUser().email,
+    date: new Date(),
+    numlikes: 0,
+    type: typePost.value,
+  };
+  addPost(objectPost)
     .then((docRef) => {
       window.location.hash = '#/home';
       newPost.value = '';
@@ -23,12 +34,13 @@ export const addDataPost = (event) => {
       console.error('Error adding document: ', error);
     });
 };
-
+  // eventShowPostToChange  ?
 export const eventShowPostToChange = (event) => {
   event.preventDefault();
   const btnShowPost = event.target;
   const newTextPost = btnShowPost.closest('.card-post').querySelector('#text-post');
   const userId = btnShowPost.closest('.card-post').querySelector('.user-post').id;
+  console.log(userId);
   const btnSave = btnShowPost.closest('.card-post').querySelector('.btn-save-change');
   const btnContSave = btnShowPost.closest('.card-post').querySelector('.update-post');
   if (currentUser().uid === userId) {
@@ -38,7 +50,7 @@ export const eventShowPostToChange = (event) => {
     btnSave.classList.remove('hidden');
   }
 };
-
+// modifica
 export const eventUpdatePost = (event) => {
   event.preventDefault();
   const btnUpdate = event.target;
@@ -68,42 +80,36 @@ export const eventDeletePost = (event) => {
   const postId = btnUpdate.closest('.card-post').id;
   const userId = btnUpdate.closest('.card-post').querySelector('.user-post').id;
   if (currentUser().uid === userId) {
-    deletePost(postId)
-      .then((doc) => {
-        console.log('Documento eliminado satisfactoriamente', doc);
-      })
-      .catch((error) => {
-        console.log(error);
-      });
+    if (window.confirm('Estas seguro que deseas eliminar este post')) {
+      deletePost(postId)
+        .then((doc) => {
+          console.log('Documento eliminado satisfactoriamente', doc);
+        })
+        .catch((error) => {
+          console.log(error);
+        });
+    }
+  } else {
+    alert('No puedes eliminar este post');
   }
 };
 
-export const eventAddComment = (event) => {
-  event.preventDefault();
-  const btnAddComment = event.target;
-  const postId = btnAddComment.closest('.card-post').id;
-  const comment = btnAddComment.closest('.card-post').querySelector('#text-comment');
-  const objComment = {
-    idPostComment: postId,
-    textComment: comment.value,
-    user: currentUser().displayName,
-  };
-  addComment(objComment)
-    .then((doc) => {
-      console.log('comentario agregado exitosamente', doc);
-    })
-    .catch((error) => {
-      console.log(error);
-    });
-};
-
-export const eventCountLikes = (event) => {
+export const addAndDeleteLikes = (event) => {
   event.preventDefault();
   const btnLike = event.target;
   const idPost = btnLike.closest('.card-post').id;
-  countLikes(idPost)
-    .then((doc) => {
-      console.log(doc);
+  getUserLike(idPost, currentUser().uid)
+    .then((likes) => {
+      console.log(likes);
+      if (!likes.empty) {
+        console.log('el usuario ya no puede dar like');
+        likes.forEach((doclike) => {
+          deleteLikes(doclike.id, idPost, currentUser().user);
+        });
+      } else {
+        console.log('el usuario puede dar like');
+        addLikes(idPost, currentUser().uid, currentUser().displayName);
+      }
     })
     .catch((error) => {
       console.log(error);
